Split alertrulesdetailpage out into separate file
