+++ web/src/components/alerting.tsx
@@ -8,23 +8,11 @@ import {
   RowProps,
   Rule,
   TableColumn,
-  Timestamp,
   useActiveNamespace,
   useListPageFilter,
   useResolvedExtensions,
   VirtualizedTable,
 } from '@openshift-console/dynamic-plugin-sdk';
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  CodeBlock,
-  CodeBlockCode,
-  DropdownItem,
-  Toolbar,
-  ToolbarContent,
-  ToolbarGroup,
-  ToolbarItem,
-} from '@patternfly/react-core';
 import { sortable } from '@patternfly/react-table';
 import classNames from 'classnames';
 import * as _ from 'lodash-es';
@@ -32,32 +20,19 @@ import * as React from 'react';
 import { Helmet } from 'react-helmet';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
-import { Link, Redirect, Route, RouteComponentProps, Switch, withRouter } from 'react-router-dom';
-
-// TODO: These will be available in future versions of the plugin SDK
-import { formatPrometheusDuration } from './console/utils/datetime';
+import { Link, Redirect, Route, RouteComponentProps, Switch } from 'react-router-dom';
 
 import { withFallback } from './console/console-shared/error/error-boundary';
-import {
-  AlertingRuleChartExtension,
-  AlertingRulesSourceExtension,
-  isAlertingRuleChart,
-  isAlertingRulesSource,
-} from './console/extensions/alerts';
-import { SectionHeading } from './console/utils/headings';
-import { ExternalLink } from './console/utils/link';
+import { AlertingRulesSourceExtension, isAlertingRulesSource } from './console/extensions/alerts';
 import { getAllQueryArguments } from './console/utils/router';
-import { EmptyBox, StatusBox } from './console/utils/status-box';
+import { EmptyBox } from './console/utils/status-box';
 
 import MonitoringDashboardsPage from './dashboards';
-import KebabDropdown from './kebab-dropdown';
-import { Labels } from './labels';
-import { QueryBrowserPage, ToggleGraph } from './metrics';
+import { QueryBrowserPage } from './metrics';
 import { CreateSilence, EditSilence } from './silence-form';
 import { TargetsUI } from './targets';
 import { Alerts, AlertSource } from './types';
 import {
-  alertDescription,
   alertingRuleStateOrder,
   alertSeverityOrder,
   fuzzyCaseInsensitive,
@@ -68,30 +43,20 @@ import './_monitoring.scss';
 import {
   getAlertRulesUrl,
   getAlertsUrl,
-  getAlertUrl,
   getIncidentsUrl,
-  getNewSilenceAlertUrl,
   getObserveState,
-  getQueryBrowserUrl,
   getRuleUrl,
   getSilencesUrl,
   usePerspective,
 } from './hooks/usePerspective';
 import {
   alertingRuleSource,
-  AlertState,
   AlertStateIcon,
   getAdditionalSources,
   getAlertStateKey,
-  getSourceKey,
-  Graph,
   MonitoringResourceIcon,
-  PopoverField,
   Severity,
-  SeverityBadge,
-  SeverityHelp,
   SilencesNotLoadedWarning,
-  SourceHelp,
 } from './alerting/AlertUtils';
 import AlertsPage from './alerting/AlertsPage';
 import AlertsDetailsPage from './alerting/AlertsDetailPage';
@@ -101,6 +66,7 @@ import { MonitoringState } from '../reducers/observe';
 import SilencesPage from './alerting/SilencesPage';
 import SilencesDetailsPage from './alerting/SilencesDetailPage';
 import { useFeatures } from './hooks/useFeatures';
+import AlertRulesDetailsPage from './alerting/AlertRulesDetailsPage';
 
 const StateCounts: React.FC<{ alerts: PrometheusAlert[] }> = ({ alerts }) => {
   const { t } = useTranslation(process.env.I18N_NAMESPACE);
@@ -120,292 +86,6 @@ const StateCounts: React.FC<{ alerts: PrometheusAlert[] }> = ({ alerts }) => {
     </>
   );
 };
-
-// Renders Prometheus template text and highlights any {{ ... }} tags that it contains
-const PrometheusTemplate = ({ text }) => (
-  <>
-    {text?.split(/(\{\{[^{}]*\}\})/)?.map((part: string, i: number) =>
-      part.match(/^\{\{[^{}]*\}\}$/) ? (
-        <code className="co-code prometheus-template-tag" key={i}>
-          {part}
-        </code>
-      ) : (
-        part
-      ),
-    )}
-  </>
-);
-
-type ActiveAlertsProps = RouteComponentProps & {
-  alerts: PrometheusAlert[];
-  namespace: string;
-  ruleID: string;
-};
-
-const ActiveAlerts_: React.FC<ActiveAlertsProps> = ({ alerts, history, namespace, ruleID }) => {
-  const { t } = useTranslation(process.env.I18N_NAMESPACE);
-  const { perspective } = usePerspective();
-
-  return (
-    <div className="co-m-table-grid co-m-table-grid--bordered">
-      <div className="row co-m-table-grid__head">
-        <div className="col-xs-6">{t('Description')}</div>
-        <div className="col-sm-2 hidden-xs">{t('Active since')}</div>
-        <div className="col-sm-2 col-xs-3">{t('State')}</div>
-        <div className="col-sm-2 col-xs-3">{t('Value')}</div>
-      </div>
-      <div className="co-m-table-grid__body">
-        {_.sortBy<PrometheusAlert>(alerts, alertDescription).map((a, i) => (
-          <div className="row co-resource-list__item" key={i}>
-            <div className="col-xs-6">
-              <Link
-                className="co-resource-item"
-                data-test="active-alerts"
-                to={getAlertUrl(perspective, a, ruleID, namespace)}
-              >
-                {alertDescription(a)}
-              </Link>
-            </div>
-            <div className="col-sm-2 hidden-xs">
-              <Timestamp timestamp={a.activeAt} />
-            </div>
-            <div className="col-sm-2 col-xs-3">
-              <AlertState state={a.state} />
-            </div>
-            <div className="col-sm-2 col-xs-3 co-truncate">{a.value}</div>
-            {a.state !== AlertStates.Silenced && (
-              <div className="dropdown-kebab-pf">
-                <KebabDropdown
-                  dropdownItems={[
-                    <DropdownItem
-                      component="button"
-                      key="silence"
-                      onClick={() => history.push(getNewSilenceAlertUrl(perspective, a))}
-                    >
-                      {t('Silence alert')}
-                    </DropdownItem>,
-                  ]}
-                />
-              </div>
-            )}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-const ActiveAlerts = withRouter(ActiveAlerts_);
-
-type AlertRulesDetailsPageProps = RouteComponentProps<{ id: string; ns?: string }>;
-
-const AlertRulesDetailsPage_: React.FC<AlertRulesDetailsPageProps> = ({ match }) => {
-  const { t } = useTranslation(process.env.I18N_NAMESPACE);
-
-  const { rulesKey, alertsKey, perspective } = usePerspective();
-  const namespace = match.params?.ns;
-
-  const rules: Rule[] = useSelector((state: MonitoringState) =>
-    getObserveState(perspective, state)?.get(rulesKey),
-  );
-  const rule = _.find(rules, { id: _.get(match, 'params.id') });
-
-  const { loaded, loadError }: Alerts = useSelector(
-    (state: MonitoringState) => getObserveState(perspective, state)?.get(alertsKey) || {},
-  );
-
-  const sourceId = rule?.sourceId;
-
-  // Load alert metrics chart from plugin
-  const [resolvedAlertsChartExtensions] =
-    useResolvedExtensions<AlertingRuleChartExtension>(isAlertingRuleChart);
-  const alertChartExtensions = resolvedAlertsChartExtensions
-    .filter((extension) => extension.properties.sourceId === sourceId)
-    .map((extension) => extension.properties.chart);
-
-  const AlertChart = alertChartExtensions[0];
-
-  const formatSeriesTitle = (alertLabels) => {
-    const nameLabel = alertLabels.__name__ ?? '';
-    const otherLabels = _.omit(alertLabels, '__name__');
-    return `${nameLabel}{${_.map(otherLabels, (v, k) => `${k}="${v}"`).join(',')}}`;
-  };
-
-  // eslint-disable-next-line camelcase
-  const runbookURL = rule?.annotations?.runbook_url;
-
-  return (
-    <>
-      <Helmet>
-        <title>{t('{{name}} details', { name: rule?.name || RuleResource.label })}</title>
-      </Helmet>
-      <StatusBox data={rule} label={RuleResource.label} loaded={loaded} loadError={loadError}>
-        <div className="pf-c-page__main-breadcrumb">
-          <Breadcrumb className="monitoring-breadcrumbs">
-            {perspective === 'dev' && (
-              <BreadcrumbItem>
-                <Link className="pf-c-breadcrumb__link" to={getAlertsUrl(perspective, namespace)}>
-                  {t('Alerts')}
-                </Link>
-              </BreadcrumbItem>
-            )}
-            {perspective !== 'dev' && (
-              <BreadcrumbItem>
-                <Link className="pf-c-breadcrumb__link" to={getAlertRulesUrl(perspective)}>
-                  {t('Alerting rules')}
-                </Link>
-              </BreadcrumbItem>
-            )}
-            <BreadcrumbItem isActive>{t('Alerting rule details')}</BreadcrumbItem>
-          </Breadcrumb>
-        </div>
-        <div className="co-m-nav-title co-m-nav-title--detail co-m-nav-title--breadcrumbs">
-          <h1 className="co-m-pane__heading">
-            <div data-test="resource-title" className="co-resource-item">
-              <MonitoringResourceIcon className="co-m-resource-icon--lg" resource={RuleResource} />
-              {rule?.name}
-              <SeverityBadge severity={rule?.labels?.severity} />
-            </div>
-          </h1>
-        </div>
-        <div className="co-m-pane__body">
-          <div className="monitoring-heading">
-            <SectionHeading text={t('Alerting rule details')} />
-          </div>
-          <div className="co-m-pane__body-group">
-            <div className="row">
-              <div className="col-sm-6">
-                <dl className="co-m-pane__details">
-                  <dt>{t('Name')}</dt>
-                  <dd>{rule?.name}</dd>
-                  <dt>
-                    <PopoverField bodyContent={<SeverityHelp />} label={t('Severity')} />
-                  </dt>
-                  <dd>
-                    <Severity severity={rule?.labels?.severity} />
-                  </dd>
-                  {rule?.annotations?.description && (
-                    <>
-                      <dt>{t('Description')}</dt>
-                      <dd>
-                        <PrometheusTemplate text={rule.annotations.description} />
-                      </dd>
-                    </>
-                  )}
-                  {rule?.annotations?.summary && (
-                    <>
-                      <dt>{t('Summary')}</dt>
-                      <dd>{rule.annotations.summary}</dd>
-                    </>
-                  )}
-                  {rule?.annotations?.message && (
-                    <>
-                      <dt>{t('Message')}</dt>
-                      <dd>
-                        <PrometheusTemplate text={rule.annotations.message} />
-                      </dd>
-                    </>
-                  )}
-                  {runbookURL && (
-                    <>
-                      <dt>{t('Runbook')}</dt>
-                      <dd>
-                        <ExternalLink href={runbookURL} text={runbookURL} />
-                      </dd>
-                    </>
-                  )}
-                </dl>
-              </div>
-              <div className="col-sm-6">
-                <dl className="co-m-pane__details">
-                  <dt>
-                    <PopoverField bodyContent={<SourceHelp />} label={t('Source')} />
-                  </dt>
-                  <dd>{rule && getSourceKey(_.startCase(alertingRuleSource(rule)), t)}</dd>
-                  {_.isInteger(rule?.duration) && (
-                    <>
-                      <dt>{t('For')}</dt>
-                      <dd>
-                        {rule.duration === 0 ? '-' : formatPrometheusDuration(rule.duration * 1000)}
-                      </dd>
-                    </>
-                  )}
-                  <dt>{t('Expression')}</dt>
-                  <dd>
-                    {/* display a link only if its a metrics based alert */}
-                    {!sourceId || sourceId === 'prometheus' ? (
-                      <Link to={getQueryBrowserUrl(perspective, rule?.query, namespace)}>
-                        <CodeBlock>
-                          <CodeBlockCode>{rule?.query}</CodeBlockCode>
-                        </CodeBlock>
-                      </Link>
-                    ) : (
-                      <CodeBlock>
-                        <CodeBlockCode>{rule?.query}</CodeBlockCode>
-                      </CodeBlock>
-                    )}
-                  </dd>
-                </dl>
-              </div>
-            </div>
-          </div>
-          <div className="co-m-pane__body-group">
-            <div className="row">
-              <div className="col-xs-12">
-                <dl className="co-m-pane__details">
-                  <dt>{t('Labels')}</dt>
-                  <dd>
-                    <Labels kind="alertrule" labels={rule?.labels} />
-                  </dd>
-                </dl>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="co-m-pane__body">
-          <div className="co-m-pane__body-group">
-            <Toolbar className="monitoring-alert-detail-toolbar">
-              <ToolbarContent>
-                <ToolbarItem variant="label">
-                  <SectionHeading text={t('Active alerts')} />
-                </ToolbarItem>
-                <ToolbarGroup align={{ default: 'alignRight' }}>
-                  <ToolbarItem>
-                    <ToggleGraph />
-                  </ToolbarItem>
-                </ToolbarGroup>
-              </ToolbarContent>
-            </Toolbar>
-            <div className="row">
-              <div className="col-sm-12">
-                {!sourceId || sourceId === 'prometheus' ? (
-                  <Graph
-                    formatSeriesTitle={formatSeriesTitle}
-                    namespace={namespace}
-                    query={rule?.query}
-                    ruleDuration={rule?.duration}
-                    showLegend
-                  />
-                ) : AlertChart ? (
-                  <AlertChart rule={rule} />
-                ) : null}
-              </div>
-            </div>
-            <div className="row">
-              <div className="col-xs-12">
-                {_.isEmpty(rule?.alerts) ? (
-                  <div className="pf-u-text-align-center">{t('None found')}</div>
-                ) : (
-                  <ActiveAlerts alerts={rule.alerts} ruleID={rule?.id} namespace={namespace} />
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-      </StatusBox>
-    </>
-  );
-};
-export const AlertRulesDetailsPage = withFallback(AlertRulesDetailsPage_);
 
 export const severityRowFilter = (t): RowFilter => ({
   filter: (filter, alert: Alert) =>
